Indicators layer refactoring. Part 2
