+++ project/OsEngine/Indicators/Aindicator.cs
@@ -846,81 +846,6 @@ namespace OsEngine.Indicators
         #endregion
     }
 
-    public class IndicatorDataSeries
-    {
-        public IndicatorDataSeries(Color color, string name, IndicatorChartPaintType paintType, bool isPaint)
-        {
-            Name = name;
-            Color = color;
-            ChartPaintType = paintType;
-            IsPaint = isPaint;
-        }
-
-        public string Name;
-
-        public Color Color;
-
-        public IndicatorChartPaintType ChartPaintType;
-
-        public bool IsPaint;
-
-        public string NameSeries;
-
-        public bool CanReBuildHistoricalValues;
-
-        public List<decimal> Values = new List<decimal>();
-
-        public decimal Last
-        {
-            get
-            {
-                if (Values.Count == 0)
-                {
-                    return 0;
-                }
-
-                return Values[Values.Count - 1];
-            }
-        }
-
-        public string GetSaveStr()
-        {
-            string result = "";
-
-            result += Name + "&";
-
-            result += Color.ToArgb() + "&";
-
-            result += ChartPaintType + "&";
-
-            result += IsPaint + "&";
-
-            return result;
-        }
-
-        public void LoadFromStr(string[] array)
-        {
-            Name = array[0];
-
-            Color = Color.FromArgb(Convert.ToInt32(array[1]));
-
-            Enum.TryParse(array[2], out ChartPaintType);
-
-            IsPaint = Convert.ToBoolean(array[3]);
-        }
-
-        public void Clear()
-        {
-            Values.Clear();
-        }
-
-        public void Delete()
-        {
-            Values.Clear();
-            Values = null;
-        }
-    }
-
     public enum IndicatorState
     {
         Configure,
