+++ project/OsEngine/Indicators/Samples/Sample2IndicatorDataSeries.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using OsEngine.Entity;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -6,7 +7,21 @@ using System.Threading.Tasks;
 
 namespace OsEngine.Indicators.Samples
 {
-    internal class Sample2IndicatorDataSeries
+    [Indicator("Sample2IndicatorDataSeries")]
+    public class Sample2IndicatorDataSeries : Aindicator
     {
+        public override void OnStateChange(IndicatorState state)
+        {
+            
+
+
+        }
+
+        public override void OnProcess(List<Candle> source, int index)
+        {
+            
+
+
+        }
     }
 }
